+++ CHANGELOG.md
@@ -1,5 +1,12 @@
+## 0.0.3
+* improve docs
+* improve code style
+* add emptyFen
+* add ChessItem.blank
+* add logger controll
+
 ## 0.0.2
-* add comments
+* add docs
 * improve readme
 
 ## 0.0.1